Remove provider runtime and its dependencies
