+++ package.json
@@ -160,8 +160,6 @@
     "@autometrics/typescript-plugin": "^0.5.3",
     "@floating-ui/core": "^1.2.6",
     "@floating-ui/react": "^0.24.2",
-    "@msgpack/msgpack": "^3.0.0-beta2",
-    "byte-base64": "^1.1.0",
     "fiberplane-charts": "git+https://git@github.com/fiberplane/fiberplane.git#workspace=fiberplane-charts",
     "fiberplane-prometheus-query": "https://git@github.com/fiberplane/fiberplane.git#workspace=fiberplane-prometheus-query",
     "framer-motion": "^10.12.16",
