+++ src/prometheus.ts
@@ -1,106 +1,20 @@
-import * as fs from "fs/promises";
-import * as path from "path";
-import type { TimeRange } from "fiberplane-charts";
 import * as vscode from "vscode";
-
-import type {
-  AutoSuggestRequest,
-  Suggestion,
-} from "./providerRuntime/suggestions";
-import type { Blob } from "./providerRuntime/types";
-import { createRuntime } from "./providerRuntime";
-import { encodeQueryData } from "./providerRuntime/queryData";
-import { formatTimeRange, uniqBy } from "./utils";
-import { fromQueryData, parseBlob } from "./providerRuntime/blobs";
-import { imports } from "./providerRuntime/imports";
 import {
-  INSTANTS_MIME_TYPE,
-  INSTANTS_QUERY_TYPE,
-  SUGGESTIONS_MIME_TYPE,
-  SUGGESTIONS_QUERY_TYPE,
-  TIMESERIES_MIME_TYPE,
-  TIMESERIES_QUERY_TYPE,
-} from "./providerRuntime/constants";
-import { matchesMimeTypeWithEncoding } from "./providerRuntime/matchesMimeTypes";
-import { unwrap } from "./providerRuntime/unwrap";
+  getStepFromTimeRange,
+  roundToGrid,
+  metricEntryToTimeseries,
+} from "fiberplane-prometheus-query";
+import type { TimeRange, Timeseries } from "fiberplane-prometheus-query";
+import fetch from "node-fetch";
+
+import { uniqBy } from "./utils";
 
 export type FunctionMetric = { moduleName: string; functionName: string };
 
-/**
- * A single data-point in time, with meta-data about the metric it was taken
- * from.
- */
-export type Instant = {
-  name: string;
-  labels: Record<string, string>;
-  metric: Metric;
-};
+export type Prometheus = ReturnType<typeof getPrometheusClient>;
 
-/**
- * A single metric value.
- *
- * Metric values are taken at a specific timestamp and contain a floating-point
- * value as well as OpenTelemetry metadata.
- */
-export type Metric = {
-  time: Timestamp;
-  value: number;
-} & OtelMetadata;
-
-/**
- * Metadata following the OpenTelemetry metadata spec.
- */
-export type OtelMetadata = {
-  // rome-ignore lint/suspicious/noExplicitAny: generated type
-  attributes: Record<string, any>;
-  // rome-ignore lint/suspicious/noExplicitAny: generated type
-  resource: Record<string, any>;
-  traceId?: OtelTraceId;
-  spanId?: OtelSpanId;
-};
-
-/**
- * SeverityNumber, as specified by OpenTelemetry:
- *  https://github.com/open-telemetry/opentelemetry-specification/blob/main/specification/logs/data-model.md#field-severitynumber
- */
-export type OtelSeverityNumber = number;
-
-/**
- * Span ID, as specified by OpenTelemetry:
- *  https://opentelemetry.io/docs/reference/specification/overview/#spancontext
- */
-export type OtelSpanId = Uint8Array;
-
-/**
- * Trace ID, as specified by OpenTelemetry:
- *  https://opentelemetry.io/docs/reference/specification/overview/#spancontext
- */
-export type OtelTraceId = Uint8Array;
-
-/**
- * A series of metrics over time, with metadata.
- */
-export type Timeseries = {
-  name: string;
-  labels: Record<string, string>;
-  metrics: Array<Metric>;
-
-  /**
-   * Whether the series should be rendered. Can be toggled by the user.
-   */
-  visible: boolean;
-} & OtelMetadata;
-
-export type Timestamp = string;
-
-export type Prometheus = Awaited<ReturnType<typeof loadPrometheusProvider>>;
-
-export async function loadPrometheusProvider(prometheusUrl: string) {
+export function getPrometheusClient(prometheusUrl: string) {
   let currentUrl = prometheusUrl;
-  const bundle = await fs.readFile(
-    path.join(__dirname, "..", "wasm", "prometheus.wasm"),
-  );
-  const provider = await createRuntime(bundle, imports);
 
   const _onDidChangeConfig: vscode.EventEmitter<void> =
     new vscode.EventEmitter<void>();
@@ -124,78 +38,26 @@ export async function loadPrometheusProvider(prometheusUrl: string) {
   }
 
   /**
+   * TODO: Implement this.
+   *
    * Fetches Autometrics function names tracked in this Prometheus instance.
    */
   async function fetchFunctions(): Promise<Array<FunctionMetric>> {
-    const blob = await _invoke(INSTANTS_QUERY_TYPE, {
-      query: "function_calls_count",
-    });
-    if (!matchesMimeTypeWithEncoding(blob.mimeType, INSTANTS_MIME_TYPE)) {
-      throw new Error(`Unexpected MIME type: ${blob.mimeType}`);
-    }
-
-    const instants = parseBlob(blob) as Array<Instant>;
-    return uniqBy(
-      instants.map((instant) => ({
-        moduleName: instant.labels.module,
-        functionName: instant.labels.function,
-      })),
-      ({ moduleName, functionName }) => `${moduleName}::${functionName}`,
-    );
+    return queryAutometricsSeries({ baseUrl: getUrl() });
   }
 
   /**
    * Fetches the names of all metrics tracked in this Prometheus instance.
    */
   async function fetchMetricNames(): Promise<Array<string>> {
-    const queryData: AutoSuggestRequest = {
-      query_type: TIMESERIES_QUERY_TYPE,
-      query: "",
-      field: "query",
-    };
-
-    const blob = await _invoke(SUGGESTIONS_QUERY_TYPE, queryData);
-    if (!matchesMimeTypeWithEncoding(blob.mimeType, SUGGESTIONS_MIME_TYPE)) {
-      throw new Error(`Unexpected MIME type: ${blob.mimeType}`);
-    }
-
-    const suggestions = parseBlob(blob) as Array<Suggestion>;
-    return suggestions
-      .filter((suggestion) => suggestion.description !== "Function")
-      .map((suggestion) => suggestion.text);
+    return queryMetricNames({ baseUrl: getUrl() });
   }
 
   async function fetchTimeseries(
     query: string,
     timeRange: TimeRange,
   ): Promise<Array<Timeseries>> {
-    const queryData = {
-      query,
-      time_range: formatTimeRange(timeRange),
-    };
-
-    const blob = await _invoke(TIMESERIES_QUERY_TYPE, queryData);
-    if (!matchesMimeTypeWithEncoding(blob.mimeType, TIMESERIES_MIME_TYPE)) {
-      throw new Error(`Unexpected MIME type: ${blob.mimeType}`);
-    }
-
-    return parseBlob(blob) as Array<Timeseries>;
-  }
-
-  function _invoke(
-    queryType: string,
-    queryData: Record<string, string>,
-  ): Promise<Blob> {
-    const invoke = provider.invoke2;
-    if (!invoke) {
-      throw new Error("invoke2() not loaded");
-    }
-
-    return invoke({
-      config: { url: currentUrl },
-      queryData: fromQueryData(encodeQueryData(queryData)),
-      queryType,
-    }).then(unwrap);
+    return queryRange(query, timeRange, { baseUrl: getUrl() });
   }
 
   return {
@@ -206,4 +68,129 @@ export async function loadPrometheusProvider(prometheusUrl: string) {
     setUrl,
     getUrl,
   };
+
+  /**
+   * Fetches the timeseries data for a given query and time range.
+   *
+   * NOTE - This is a copy of the `querySeries` function from fiberplane-prometheus-query.
+   *        I needed to copy it here to fix the URL param
+   *        Once `querySeries` is fixed, we can remove this function.
+   */
+  async function queryRange(
+    query: string,
+    timeRange: TimeRange,
+    { baseUrl }: { baseUrl: string },
+  ) {
+    const [stepParam, stepSeconds] = getStepFromTimeRange(timeRange);
+
+    const params = new URLSearchParams();
+
+    params.append("query", query);
+    params.append(
+      "start",
+      roundToGrid(timeRange.from, stepSeconds, Math.floor),
+    );
+    params.append("end", roundToGrid(timeRange.to, stepSeconds, Math.ceil));
+    params.append("step", stepParam);
+
+    const url = `${baseUrl}/api/v1/query_range?${params.toString()}`;
+    const response = await fetch(url, { mode: "cors" });
+
+    if (!response.ok) {
+      throw new Error("Error fetching prometheus data");
+    }
+
+    const jsonResponse = await response.json();
+    if (!isObject(jsonResponse)) {
+      throw new Error("Unexpected response from Prometheus");
+    }
+
+    const { data } = jsonResponse;
+    if (!isObject(data)) {
+      throw new Error("Invalid or missing data in Prometheus response");
+    }
+
+    const { result } = data;
+    if (!Array.isArray(result)) {
+      throw new Error("Invalid or missing results in Prometheus response");
+    }
+
+    return result.map(metricEntryToTimeseries);
+  }
+}
+
+/**
+ * Queries Prometheus for all autometrics-matching series
+ */
+async function queryAutometricsSeries({ baseUrl }: { baseUrl: string }) {
+  // Search for all function_calls series, and only include results that have both a function and module label
+  const params = new URLSearchParams();
+  params.append(
+    "match[]",
+    '{__name__=~"function_calls(_count)?(_total)?", function!="", module!=""}',
+  );
+
+  const url = `${baseUrl}/api/v1/series?${params.toString()}`;
+  const response = await fetch(url, { mode: "cors" });
+
+  if (!response.ok) {
+    throw new Error("Error fetching prometheus data");
+  }
+
+  const jsonResponse = await response.json();
+  if (!isObject(jsonResponse)) {
+    throw new Error("Unexpected response from Prometheus");
+  }
+
+  const { data } = jsonResponse;
+  if (!Array.isArray(data)) {
+    throw new Error("Invalid or missing data in Prometheus response");
+  }
+
+  return uniqBy(
+    data
+      .filter(
+        (series) => isString(series?.function) && isString(series?.module),
+      )
+      .map((series) => ({
+        functionName: String(series.function),
+        moduleName: String(series.module),
+      })),
+    ({ moduleName, functionName }) => `${moduleName}::${functionName}`,
+  );
+}
+
+/**
+ * Queries Prometheus for all autometrics-matching series
+ */
+async function queryMetricNames({ baseUrl }: { baseUrl: string }) {
+  const url = `${baseUrl}/api/v1/label/__name__/values`;
+  const response = await fetch(url, { mode: "cors" });
+
+  if (!response.ok) {
+    throw new Error("Error fetching prometheus data");
+  }
+
+  const jsonResponse = await response.json();
+  if (!isObject(jsonResponse)) {
+    throw new Error("Unexpected response from Prometheus");
+  }
+
+  const { data } = jsonResponse;
+
+  if (!Array.isArray(data) || !data.every(isString)) {
+    throw new Error("Invalid or missing data in Prometheus response");
+  }
+
+  return data;
+}
+
+function isObject(
+  maybeObject: unknown,
+): maybeObject is Record<string, unknown> {
+  return typeof maybeObject === "object" && maybeObject != null;
+}
+
+function isString(maybeString: unknown): maybeString is string {
+  return typeof maybeString === "string";
 }
