--- src/providerRuntime/constants.ts
@@ -1,12 +1,0 @@
-export const FORM_ENCODED_MIME_TYPE = "application/x-www-form-urlencoded";
-
-export const INSTANTS_MIME_TYPE = "application/vnd.fiberplane.instants";
-export const INSTANTS_QUERY_TYPE = "x-instants";
-
-export const SUGGESTIONS_MIME_TYPE = "application/vnd.fiberplane.suggestions";
-export const SUGGESTIONS_QUERY_TYPE = "suggestions";
-
-export const SUPPORTED_ENCODINGS = ["json", "msgpack"];
-
-export const TIMESERIES_MIME_TYPE = "application/vnd.fiberplane.timeseries";
-export const TIMESERIES_QUERY_TYPE = "timeseries";
